+++ src/gmprocess/utils/ground_motion_models.py
@@ -1,6 +1,3 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
 from importlib import import_module
 from ruamel.yaml import YAML